+++ app/components/doi-creator.js
@@ -1,4 +1,4 @@
-import PersonBaseComponent from '../components/person';
+import PersonBaseComponent from './person-base-component';
 
 export default PersonBaseComponent.extend({
   actions: {
